+++ packages/nonstop_cli/bricks/flutter_package_for_mono_repo/hooks/post_gen.dart
@@ -1,0 +1,15 @@
+import 'dart:async';
+
+import 'package:mason/mason.dart';
+
+import 'commands/flutter_package_create_command.dart';
+
+Future<void> run(HookContext context) async {
+  final commands = [
+    FlutterPackageCreateCommand(),
+  ];
+
+  for (final command in commands) {
+    await command.run(context);
+  }
+}